move dynamic sampling into own crate
