+++ relay-sampling/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "relay-sampling"
+authors = ["Sentry <oss@sentry.io>"]
+description = "Dynamic sampling functionality for Relay"
+homepage = "https://getsentry.github.io/relay/"
+repository = "https://github.com/getsentry/relay"
+version = "21.2.0"
+edition = "2018"
+license-file = "../LICENSE"
+publish = false
+
+[dependencies]
+relay-common = { path = "../relay-common" }
+relay-general = { path = "../relay-general" }
+relay-log = { path = "../relay-log" }
+serde = { version = "1.0.114", features = ["derive"] }
+serde_json = "1.0.55"
+relay-filter = { path = "../relay-filter" }
+rand= "0.6.5"
+rand_pcg="0.1.2"
+unicase="2.6.0"
+
+
+[dev-dependencies]
+insta = {version="0.16.0", features=["ron"]}
